Put Criteo's TemporaryFile in separate namespace
